--- a/ggml/src/ggml.c
+++ b/ggml/src/ggml.c
@@ -1619,7 +1619,6 @@ static const ggml_type_traits_t type_traits[GGML_TYPE_COUNT] = {
         .vec_dot                  = vec_dot_iq4_kt_q8_k,
 #ifdef __ARM_NEON
         .vec_dot_type             = GGML_TYPE_F16,
-        //.vec_dot_type             = GGML_TYPE_F32,
 #else
         .vec_dot_type             = GGML_TYPE_F32,
 #endif