--- a/ggml/src/iqk/iqk_gemm_ktquants.h
+++ b/ggml/src/iqk/iqk_gemm_ktquants.h
@@ -8,6 +8,6 @@
 
 bool iqk_set_kernels_ktquants(int ne00, int typeA, int typeB, std::array<mul_mat_t, IQK_MAX_NY>& kernels, mul_mat_t& func16);
 
-void iqk_dequantize_iq4_kt(int n, const void * vx, size_t bx, float * y, size_t stride_y, int nrc_x);
+bool iqk_dequantize_ktquants(int type, int n, const void * vx, size_t bx, float * y, size_t stride_y, int nrc_x);
 
 #endif